refactor: extract core TPM logic to src/tpm_utils.py
